Values saved to DB as integer/real instead of text. User can choose units for daily entry and this is reflected in the listView. Count has no unit. Number of calories is calculated when entry value changes or the units are toggled

+++ app/src/main/java/com/example/Smartscale/webAPIFood.java
@@ -2,8 +2,8 @@ package com.example.Smartscale;
 
 public class webAPIFood {
         String foodName;
-        double calories;
-        double mass;
+        int calories;
+        int mass;
 
 public String toString()
 {
